requirements.txt
dj-database-url==0.4.1
Django==1.10.6
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
requests==2.9.1
pytz
pipeline
django-cors-headers
djangorestframework
